// src/types/resources/terminal.types.ts
export type TerminalTheme = {
  name: string;
  colors: Record<
    | "color0"
    | "color1"
    | "color2"
    | "color3"
    | "color4"
    | "color5"
    | "color6"
    | "color7"
    | "color8"
    | "color9"
    | "color10"
    | "color11"
    | "color12"
    | "color13"
    | "color14"
    | "color15",
    string
  >;
};

export type TerminalInstaller = {
  slug: string;
  name: string;
  commands: string[];
  isWindowsOnly?: boolean;
};

export type TerminalResources = {
  themes: TerminalTheme[];
  terminals: TerminalInstaller[];
  universalInstaller?: string;
  xfetchLogo?: string;
};